docs: update changelog for router refactor
